+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🐛 Fixes
 
 - [10969](https://github.com/vegaprotocol/vega/issues/10969) - Ensure teams statistics are computed from team rewards.
+- [10973](https://github.com/vegaprotocol/vega/issues/10973) - Avoid entering an auction or doing mark-to-market before market entered opening auction or after it is in a terminal state. 
 
 ## 0.75.2